ragas/requirements.txt
# requirements.txt
ragas>=0.0.20
datasets>=2.15.0
pandas>=2.0.0
pyyaml>=6.0.1
torch>=2.1.0
transformers>=4.36.0
numpy>=1.24.0
tenacity>=8.2.3
python-dotenv>=1.0.0
google-generativeai>=0.3.1
aiohttp>=3.9.1  # For async support